+++ model/publishing/delivery/tasks/DeployTestEnvironments.php
@@ -1,26 +1,37 @@
 <?php
-/**  
+
+/**
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
  * as published by the Free Software Foundation; under version 2
  * of the License (non-upgradable).
- * 
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public License
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
- * 
- * Copyright (c) 2017 (original work) Open Assessment Technologies SA;
- *               
- * 
+ *
+ * Copyright (c) 2020 (original work) Open Assessment Technologies SA;
+ *
+ *
  */
+declare(strict_types=1);
+
 namespace oat\taoPublishing\model\publishing\delivery\tasks;
 
-use oat\generis\model\OntologyRdfs;
+use common_exception_Error;
+use common_exception_MissingParameter;
+use common_Logger;
+use common_report_Report;
+use core_kernel_classes_Resource;
+use Exception;
+use GuzzleHttp\Psr7\MultipartStream;
+use GuzzleHttp\Psr7\Request;
+use oat\generis\model\OntologyAwareTrait;
 use oat\oatbox\action\Action;
 use oat\tao\model\taskQueue\QueueDispatcherInterface;
 use oat\tao\model\taskQueue\Task\ChildTaskAwareInterface;
@@ -30,17 +41,15 @@ use oat\tao\model\taskQueue\Task\TaskAwareTrait;
 use oat\taoDeliveryRdf\controller\RestTest;
 use oat\taoDeliveryRdf\model\DeliveryAssemblyService;
 use oat\taoPublishing\model\PlatformService;
-use oat\taoPublishing\model\publishing\delivery\PublishingDeliveryService;
-use Zend\ServiceManager\ServiceLocatorAwareTrait;
+use oat\taoPublishing\model\publishing\delivery\DeliveryTestService;
 use Zend\ServiceManager\ServiceLocatorAwareInterface;
-use oat\generis\model\OntologyAwareTrait;
-use GuzzleHttp\Psr7\Request;
-use GuzzleHttp\Psr7\MultipartStream;
+use Zend\ServiceManager\ServiceLocatorAwareTrait;
 
 /**
  * Deploys a test to environments
  */
-class DeployTestEnvironments implements Action, ServiceLocatorAwareInterface, ChildTaskAwareInterface, TaskAwareInterface
+class DeployTestEnvironments implements Action, ServiceLocatorAwareInterface, ChildTaskAwareInterface,
+                                        TaskAwareInterface
 {
     use ServiceLocatorAwareTrait;
     use OntologyAwareTrait;
@@ -48,64 +57,63 @@ class DeployTestEnvironments implements Action, ServiceLocatorAwareInterface, Ch
     use ChildTaskAwareTrait;
 
     /**
-     * @param array  $params
-     * @return \common_report_Report
-     * @throws \common_exception_Error
-     * @throws \common_exception_MissingParameter
+     * @param array $params
+     *
+     * @return common_report_Report
+     *
+     * @throws common_exception_Error
+     * @throws common_exception_MissingParameter
      */
-    public function __invoke($params) {
-
+    public function __invoke($params): common_report_Report
+    {
         if (count($params) != 3) {
-            throw new \common_exception_MissingParameter();
+            throw new common_exception_MissingParameter();
         }
         $test = $this->getResource(array_shift($params));
         $envId = array_shift($params);
         $env = $this->getResource($envId);
         $deliveryId = array_shift($params);
         $delivery = $this->getResource($deliveryId);
-        \common_Logger::i('Deploying '.$test->getLabel().' to '.$env->getLabel());
-        $report = new \common_report_Report(\common_report_Report::TYPE_SUCCESS, __('Deployed %s to %s', $test->getLabel(), $env->getLabel()));
-        
+        common_Logger::i('Deploying ' . $test->getLabel() . ' to ' . $env->getLabel());
+        $report = new common_report_Report(
+            common_report_Report::TYPE_SUCCESS,
+            __('Deployed %s to %s', $test->getLabel(), $env->getLabel())
+        );
+
         $subReport = $this->compileTest($env, $test, $delivery);
         $report->add($subReport);
+
         return $report;
     }
 
     /**
-     * @param \core_kernel_classes_Resource $env
-     * @param \core_kernel_classes_Resource $test
-     * @param \core_kernel_classes_Resource $delivery
-     * @return \common_report_Report
+     * @param core_kernel_classes_Resource $env
+     * @param core_kernel_classes_Resource $test
+     * @param core_kernel_classes_Resource $delivery
+     *
+     * @return common_report_Report
      */
-    protected function compileTest(\core_kernel_classes_Resource $env, \core_kernel_classes_Resource $test, \core_kernel_classes_Resource $delivery) {
+    protected function compileTest(
+        core_kernel_classes_Resource $env,
+        core_kernel_classes_Resource $test,
+        core_kernel_classes_Resource $delivery
+    ): common_report_Report {
         try {
-            \common_Logger::d('Exporting Test '.$test->getUri().' for deployment');
-            $exporter = new \taoQtiTest_models_classes_export_TestExport();
-            $exportReport = $exporter->export([
-                'filename' => \League\Flysystem\Util::normalizePath($test->getLabel()),
-                'instances' => $test->getUri(),
-            ], \tao_helpers_File::createTempDir());
-            $packagePath = $exportReport->getData();
-            if(is_array($packagePath) && isset($packagePath['path'])){
-                $packagePath = $packagePath['path'];
-            }
-            $streamData = [[
-                'name'     => RestTest::REST_FILE_NAME,
-                'contents' => fopen($packagePath, 'rb'),
-            ], [
-                'name'     => RestTest::REST_IMPORTER_ID,
-                'contents' => 'taoQtiTest',
-            ], [
-                'name'     => RestTest::REST_DELIVERY_PARAMS,
-                'contents' => json_encode([
-                    PublishingDeliveryService::ORIGIN_DELIVERY_ID_FIELD => $delivery->getUri()
-                ])
-            ]];
+            $streamData = [
+                [
+                    'name' => RestTest::REST_FILE_NAME,
+                    'contents' => $this->getDeliveryTestService()->getTestStream($delivery),
+                ],
+                [
+                    'name' => RestTest::REST_IMPORTER_ID,
+                    'contents' => 'taoQtiTest',
+                ]
+            ];
 
             $deliveryClass = current($delivery->getTypes());
             if ($deliveryClass->getUri() != DeliveryAssemblyService::CLASS_URI) {
                 $streamData[] = [
-                    'name'     => RestTest::REST_DELIVERY_CLASS_LABEL,
+                    'name' => RestTest::REST_DELIVERY_CLASS_LABEL,
                     'contents' => $deliveryClass->getLabel()
                 ];
             }
@@ -114,7 +122,7 @@ class DeployTestEnvironments implements Action, ServiceLocatorAwareInterface, Ch
             $request = new Request('POST', '/taoDeliveryRdf/RestTest/compileDeferred');
             $request = $request->withBody($body);
 
-            \common_Logger::d('Requesting compilation of Test '.$test);
+            common_Logger::d('Requesting compilation of Test ' . $test);
 
             $response = PlatformService::singleton()->callApi($env->getUri(), $request);
             if ($response->getStatusCode() == 200) {
@@ -123,7 +131,7 @@ class DeployTestEnvironments implements Action, ServiceLocatorAwareInterface, Ch
                     $data = $content['data'];
                     $taskId = $data['reference_id'];
 
-                    /** @var QueueDispatcherInterface $queueDispatcher reference_id*/
+                    /** @var QueueDispatcherInterface $queueDispatcher reference_id */
                     $queueDispatcher = $this->getServiceLocator()->get(QueueDispatcherInterface::SERVICE_ID);
 
                     $queueDispatcher->createTask(
@@ -134,14 +142,33 @@ class DeployTestEnvironments implements Action, ServiceLocatorAwareInterface, Ch
                     );
                     $this->addChildId($taskId);
                 }
-                $report = new \common_report_Report(\common_report_Report::TYPE_SUCCESS, __('Test has been compiled as %s', $delivery->getUri()));
+                $report = new common_report_Report(
+                    common_report_Report::TYPE_SUCCESS,
+                    __('Test has been compiled as %s', $delivery->getUri())
+                );
                 $report->setData($delivery->getUri());
                 return $report;
             } else {
-                return new \common_report_Report(\common_report_Report::TYPE_ERROR, __('Failed to compile %s with message: %s', $test, $response->getBody()->getContents()));
+                return new common_report_Report(
+                    common_report_Report::TYPE_ERROR,
+                    __(
+                        'Failed to compile %s with message: %s',
+                        $test,
+                        $response->getBody()->getContents()
+                    )
+                );
             }
-        } catch (\Exception $e) {
-            return new \common_report_Report(\common_report_Report::TYPE_ERROR, __('Failed to compile %s with message: %s', $test, $e->getMessage()));
+        } catch (Exception $e) {
+            return new common_report_Report(
+                common_report_Report::TYPE_ERROR,
+                __('Failed to compile %s with message: %s', $test, $e->getMessage())
+            );
         }
+    }
+
+    private function getDeliveryTestService(): DeliveryTestService
+    {
+        /** @noinspection PhpIncompatibleReturnTypeInspection */
+        return $this->getServiceLocator()->get(DeliveryTestService::SERVICE_ID);
     }
 }